Compute the renforcement learning reward asynchronously
wip
